Preferences: Don't start the FileDisplayActivity with FLAG_ACTIVITY_CLEAR_TOP.
Since the FileDisplayActivity has its launch mode set to 'multiple' (the default) and FLAG_ACTIVITY_SINGLE_TOP is not being used, using FLAG_ACTIVITY_CLEAR_TOP will effectively finish any existing activity and then re-create it. This will cause the launch screen to be showed again. This is probably not the behaviour we want.